# ai_stock_analyzer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Stock Analyzer API
  description: API untuk mengambil data harga saham historis, indikator teknikal (VPA/Wyckoff), dan analisis/rekomendasi harian berbasis AI.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register akun pengguna baru
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
  /auth/token:
    post:
      tags:
        - Authentication
      summary: Login pengguna untuk mendapatkan JWT Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
        '401':
          description: Email atau password salah
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh Access Token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
        '401':
          description: Token invalid
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout dan revoke token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Success
...
  /analysis/{ticker}:
    post:
      tags:
        - Analysis
      summary: Jalankan AI Analysis untuk saham
      parameters:
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /analysis/{ticker}/backtest:
    post:
      tags:
        - Analysis
      summary: Jalankan Backtest Strategi
      parameters:
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                years:
                  type: integer
                target_profit_pct:
                  type: number
                stop_loss_pct:
                  type: number
      responses:
        '200':
          description: Success
...